refactor: move democracy consumer code to /x/ccv/consumer (#883)
* move utils to types

* regen pb files

* todo

* params -> consumerParams and providerParams

* mas proto changes

* consumer done

* mas

* rm prov genesis pb

* add

* mas

* mas

* progress is progress

* progresso chicken noodle soup

* un mas, calling it for today

* mas

* passes UTs

* woot

* diff tests

* fix router key bug

* fix tx commands

* Merge branch 'main' into add-go-mods

* lint fixes

* fix lint

* move the ish

* Update app.go

* Revert "Merge branch 'main' into add-go-mods"

This reverts commit 1e91d47ac8866b5c730e24e9f4e559adf2d8ab66.

* Revert "Revert "Merge branch 'main' into add-go-mods""

This reverts commit b1c4f8fc471c1cf698e2d620b2f2f80876f13a0c.
